feat: helper refactoring
