+++ src/utils/api/helpers/donationHelper.ts
@@ -1,0 +1,38 @@
+import { getConfigOrThrow } from "../../../utils/config/config";
+import {
+  DONATION_INIT_SESSION,
+  DONATION_LIST_ERROR,
+  DONATION_LIST_SUCCESS,
+} from "../../../utils/config/mixpanelDefs";
+import { ErrorsType } from "../../../utils/errors/checkErrorsModel";
+import { mixpanel } from "../../../utils/config/mixpanelHelperInit";
+
+export const getDonationEntityList = async (
+  onError: (e: string) => void,
+  onResponse: (data: any) => void
+) => {
+  mixpanel.track(DONATION_INIT_SESSION.value, {
+    EVENT_ID: DONATION_INIT_SESSION.value,
+  });
+
+  window
+    .fetch(getConfigOrThrow().CHECKOUT_DONATIONS_URL)
+    .then((response) => {
+      if (response.ok) {
+        return response.json();
+      }
+      throw new Error(ErrorsType.DONATIONLIST_ERROR);
+    })
+    .then((data) => {
+      mixpanel.track(DONATION_LIST_SUCCESS.value, {
+        EVENT_ID: DONATION_LIST_SUCCESS.value,
+      });
+      onResponse(data);
+    })
+    .catch(() => {
+      mixpanel.track(DONATION_LIST_ERROR.value, {
+        EVENT_ID: DONATION_LIST_ERROR.value,
+      });
+      onError(ErrorsType.DONATIONLIST_ERROR);
+    });
+};
