+++ src/utils/api/helpers/ecommerce/ecommercePaymentMethodsHelper.ts
@@ -1,0 +1,461 @@
+/* eslint-disable sonarjs/no-identical-functions */
+import * as E from "fp-ts/Either";
+import { pipe } from "fp-ts/lib/function";
+import * as O from "fp-ts/Option";
+import * as TE from "fp-ts/TaskEither";
+import {
+  apiPaymentEcommerceClient,
+  apiPaymentEcommerceClientV3,
+  apiPaymentEcommerceClientWithRetry,
+  apiPaymentEcommerceClientWithRetryV2,
+  apiPaymentEcommerceClientWithRetryV3,
+} from "../../../../utils/api/client";
+import { ErrorsType } from "../../../../utils/errors/checkErrorsModel";
+import { mixpanel } from "../../../../utils/config/mixpanelHelperInit";
+import {
+  NPG_INIT,
+  NPG_NET_ERR,
+  NPG_RESP_ERROR,
+  NPG_SUCCESS,
+  NPG_SVR_ERR,
+  PAYMENT_METHODS_ACCESS,
+  PAYMENT_METHODS_NET_ERROR,
+  PAYMENT_METHODS_RESP_ERROR,
+  PAYMENT_METHODS_SUCCESS,
+  PAYMENT_METHODS_SVR_ERROR,
+  PAYMENT_PSPLIST_INIT,
+  PAYMENT_PSPLIST_NET_ERR,
+  PAYMENT_PSPLIST_RESP_ERR,
+  PAYMENT_PSPLIST_SUCCESS,
+  PAYMENT_PSPLIST_SVR_ERR,
+} from "../../../../utils/config/mixpanelDefs";
+import {
+  PaymentInstrumentsType,
+  PaymentMethod,
+  PaymentMethodInfo,
+} from "../../../../features/payment/models/paymentModel";
+import { validateSessionWalletCardFormFields } from "../../../../utils/regex/validators";
+import {
+  getRptIdsFromSession,
+  getSessionItem,
+  SessionItems,
+  setSessionItem,
+} from "../../../../utils/storage/sessionStorage";
+import { NewTransactionResponse } from "../../../../../generated/definitions/payment-ecommerce/NewTransactionResponse";
+import { CalculateFeeRequest } from "../../../../../generated/definitions/payment-ecommerce-v2/CalculateFeeRequest";
+import { CreateSessionResponse } from "../../../../../generated/definitions/payment-ecommerce/CreateSessionResponse";
+import { Bundle } from "../../../../../generated/definitions/payment-ecommerce-v2/Bundle";
+import { CalculateFeeResponse } from "../../../../../generated/definitions/payment-ecommerce-v2/CalculateFeeResponse";
+
+// ->Promise<Either<string,SessionPaymentMethodResponse>>
+export const retrieveCardData = async ({
+  paymentId,
+  orderId,
+  onError,
+  onResponseSessionPaymentMethod,
+}: {
+  paymentId: string;
+  orderId: string;
+  onError: (e: string) => void;
+  onResponseSessionPaymentMethod: (r: any) => void;
+}) => {
+  const transactionId = pipe(
+    getSessionItem(SessionItems.transaction) as NewTransactionResponse,
+    O.fromNullable,
+    O.map((transaction) => transaction.transactionId),
+    O.getOrElse(() => "")
+  );
+
+  const bearerAuth = pipe(
+    getSessionItem(SessionItems.transaction) as NewTransactionResponse,
+    O.fromNullable,
+    O.chain((transaction) => O.fromNullable(transaction.authToken)),
+    O.getOrElse(() => "")
+  );
+
+  /*
+      mixpanel.track(PAYMENT_PSPLIST_INIT.value, {
+      EVENT_ID: PAYMENT_PSPLIST_INIT.value,
+     }); */
+
+  const sessionPaymentMethod = await pipe(
+    TE.tryCatch(
+      () =>
+        apiPaymentEcommerceClient.getSessionPaymentMethod({
+          bearerAuth,
+          id: paymentId,
+          orderId,
+          "x-transaction-id-from-client": transactionId,
+        }),
+      (_e) => {
+        onError(ErrorsType.CONNECTION);
+        mixpanel.track(PAYMENT_PSPLIST_NET_ERR.value, {
+          EVENT_ID: PAYMENT_PSPLIST_NET_ERR.value,
+        });
+        return E.toError;
+      }
+    ),
+    TE.fold(
+      (_r) => async () => {
+        onError(ErrorsType.SERVER);
+        mixpanel.track(PAYMENT_PSPLIST_SVR_ERR.value, {
+          EVENT_ID: PAYMENT_PSPLIST_SVR_ERR.value,
+        });
+        return {};
+      },
+      (myResExt) => async () =>
+        pipe(
+          myResExt,
+          E.fold(
+            () => [],
+            (myRes) => {
+              if (myRes?.status === 200) {
+                const sessionPaymentMethodResponse = myRes?.value;
+                setSessionItem(
+                  SessionItems.sessionPaymentMethod,
+                  sessionPaymentMethodResponse
+                );
+                setSessionItem(SessionItems.paymentMethodInfo, {
+                  title: `· · · · ${sessionPaymentMethodResponse.lastFourDigits}`,
+                  body: sessionPaymentMethodResponse.expiringDate,
+                  icon: sessionPaymentMethodResponse.brand,
+                } as PaymentMethodInfo);
+                return sessionPaymentMethodResponse;
+              } else {
+                onError(ErrorsType.GENERIC_ERROR);
+                return {};
+              }
+            }
+          )
+        )
+    )
+  )();
+
+  onResponseSessionPaymentMethod(sessionPaymentMethod);
+};
+
+// -> Promise<Either<string, BundleOptions>>
+export const calculateFees = async ({
+  paymentId,
+  bin,
+  onError,
+  onPspNotFound,
+  onResponsePsp,
+}: {
+  paymentId: string;
+  bin?: string;
+  onError: (e: string) => void;
+  onPspNotFound: () => void;
+  onResponsePsp: (r: any) => void;
+}) => {
+  const calculateFeeRequest: O.Option<CalculateFeeRequest> = pipe(
+    getSessionItem(SessionItems.transaction) as NewTransactionResponse,
+    O.fromNullable,
+    O.map((transaction) => ({
+      bin,
+      touchpoint: transaction.clientId || "CHECKOUT",
+      paymentNotices: transaction.payments.map((payment) => ({
+        paymentAmount: payment.amount,
+        primaryCreditorInstitution: payment.rptId.substring(0, 11),
+        transferList: payment.transferList.map((transfer) => ({
+          creditorInstitution: transfer.paFiscalCode,
+          digitalStamp: transfer.digitalStamp,
+          transferCategory: transfer.transferCategory,
+        })),
+      })),
+      isAllCCP: transaction.payments[0].isAllCCP,
+    }))
+  );
+
+  const transactionId = pipe(
+    getSessionItem(SessionItems.transaction) as NewTransactionResponse,
+    O.fromNullable,
+    O.map((transaction) => transaction.transactionId),
+    O.getOrElse(() => "")
+  );
+
+  const bearerAuth = pipe(
+    getSessionItem(SessionItems.transaction) as NewTransactionResponse,
+    O.fromNullable,
+    O.chain((transaction) => O.fromNullable(transaction.authToken)),
+    O.getOrElse(() => "")
+  );
+
+  mixpanel.track(PAYMENT_PSPLIST_INIT.value, {
+    EVENT_ID: PAYMENT_PSPLIST_INIT.value,
+  });
+  const MAX_OCCURENCES_AFM = 2147483647;
+  await pipe(
+    calculateFeeRequest,
+    TE.fromOption(() => {
+      onError(ErrorsType.GENERIC_ERROR);
+      return E.toError;
+    }),
+    TE.chain((calculateFeeRequest) =>
+      TE.tryCatch(
+        () =>
+          apiPaymentEcommerceClientWithRetryV2.calculateFees({
+            bearerAuth,
+            "x-transaction-id-from-client": transactionId,
+            id: paymentId,
+            maxOccurrences: MAX_OCCURENCES_AFM,
+            body: calculateFeeRequest,
+          }),
+        (_e) => {
+          onError(ErrorsType.CONNECTION);
+          mixpanel.track(PAYMENT_PSPLIST_NET_ERR.value, {
+            EVENT_ID: PAYMENT_PSPLIST_NET_ERR.value,
+          });
+          return E.toError;
+        }
+      )
+    ),
+    TE.fold(
+      (_r) => async () => {
+        onError(ErrorsType.SERVER);
+        mixpanel.track(PAYMENT_PSPLIST_SVR_ERR.value, {
+          EVENT_ID: PAYMENT_PSPLIST_SVR_ERR.value,
+        });
+        return {};
+      },
+      (myResExt) => async () =>
+        pipe(
+          myResExt,
+          E.fold(
+            () => [],
+            (myRes) => {
+              if (myRes?.status === 200) {
+                mixpanel.track(PAYMENT_PSPLIST_SUCCESS.value, {
+                  EVENT_ID: PAYMENT_PSPLIST_SUCCESS.value,
+                });
+                onResponsePsp(myRes?.value);
+              } else if (myRes?.status === 404) {
+                onPspNotFound();
+                mixpanel.track(PAYMENT_PSPLIST_RESP_ERR.value, {
+                  EVENT_ID: PAYMENT_PSPLIST_RESP_ERR.value,
+                });
+              } else {
+                onError(ErrorsType.GENERIC_ERROR);
+                mixpanel.track(PAYMENT_PSPLIST_RESP_ERR.value, {
+                  EVENT_ID: PAYMENT_PSPLIST_RESP_ERR.value,
+                });
+              }
+              return {};
+            }
+          )
+        )
+    )
+  )();
+};
+
+export const getPaymentInstruments = async (
+  query: {
+    amount: number;
+  },
+  onError: (e: string) => void,
+  onResponse: (data: Array<PaymentInstrumentsType>) => void
+) => {
+  mixpanel.track(PAYMENT_METHODS_ACCESS.value, {
+    EVENT_ID: PAYMENT_METHODS_ACCESS.value,
+  });
+  const list = await pipe(
+    TE.tryCatch(
+      () =>
+        pipe(
+          getSessionItem(SessionItems.authToken) as string,
+          O.fromNullable,
+          O.fold(
+            () => apiPaymentEcommerceClient.getAllPaymentMethods(query),
+            (bearerAuth) =>
+              apiPaymentEcommerceClientV3.getAllPaymentMethodsV3({
+                "x-rpt-id": getRptIdsFromSession(),
+                bearerAuth,
+                ...query,
+              })
+          )
+        ),
+      () => {
+        mixpanel.track(PAYMENT_METHODS_NET_ERROR.value, {
+          EVENT_ID: PAYMENT_METHODS_NET_ERROR.value,
+        });
+        onError(ErrorsType.STATUS_ERROR);
+        return E.toError;
+      }
+    ),
+    TE.fold(
+      () => async () => {
+        mixpanel.track(PAYMENT_METHODS_SVR_ERROR.value, {
+          EVENT_ID: PAYMENT_METHODS_SVR_ERROR.value,
+        });
+        onError(ErrorsType.STATUS_ERROR);
+        return [];
+      },
+      (myResExt) => async () =>
+        pipe(
+          myResExt,
+          E.fold(
+            () => {
+              mixpanel.track(PAYMENT_METHODS_RESP_ERROR.value, {
+                EVENT_ID: PAYMENT_METHODS_RESP_ERROR.value,
+              });
+              return [];
+            },
+            (myRes) => {
+              switch (myRes.status) {
+                case 200:
+                  mixpanel.track(PAYMENT_METHODS_SUCCESS.value, {
+                    EVENT_ID: PAYMENT_METHODS_SUCCESS.value,
+                  });
+                  return myRes.value.paymentMethods;
+                case 401:
+                  mixpanel.track(PAYMENT_METHODS_RESP_ERROR.value, {
+                    EVENT_ID: PAYMENT_METHODS_RESP_ERROR.value,
+                  });
+                  onError(ErrorsType.UNAUTHORIZED);
+                  return [];
+                default:
+                  mixpanel.track(PAYMENT_METHODS_RESP_ERROR.value, {
+                    EVENT_ID: PAYMENT_METHODS_RESP_ERROR.value,
+                  });
+                  return [];
+              }
+            }
+          )
+        )
+    )
+  )();
+  onResponse(list as any as Array<PaymentInstrumentsType>);
+};
+
+export const npgSessionsFields = async (
+  onError: (e: string) => void,
+  onResponse: (data: CreateSessionResponse) => void
+) =>
+  await pipe(
+    TE.tryCatch(
+      () => {
+        mixpanel.track(NPG_INIT.value, {
+          EVENT_ID: NPG_INIT.value,
+        });
+        const paymentMethodId =
+          (
+            getSessionItem(SessionItems.paymentMethod) as
+              | PaymentMethod
+              | undefined
+          )?.paymentMethodId || "";
+        const payload = {
+          id: paymentMethodId,
+          lang: localStorage.getItem("i18nextLng") ?? "it",
+        };
+        return pipe(
+          getSessionItem(SessionItems.authToken) as string,
+          O.fromNullable,
+          O.fold(
+            () =>
+              apiPaymentEcommerceClientWithRetry.createSession({
+                ...payload,
+              }),
+            (bearerAuth) =>
+              apiPaymentEcommerceClientWithRetryV3.createSessionV3({
+                "x-rpt-id": getRptIdsFromSession(),
+                bearerAuth,
+                ...payload,
+              })
+          )
+        );
+      },
+      () => {
+        mixpanel.track(NPG_NET_ERR.value, {
+          EVENT_ID: NPG_NET_ERR.value,
+        });
+        onError(NPG_NET_ERR.value);
+        return NPG_NET_ERR.value;
+      }
+    ),
+    TE.fold(
+      (err) => {
+        mixpanel.track(NPG_SVR_ERR.value, {
+          EVENT_ID: NPG_SVR_ERR.value,
+        });
+        return TE.left(err);
+      },
+      (myResExt) => async () =>
+        pipe(
+          myResExt,
+          E.fold(
+            () => {
+              mixpanel.track(NPG_RESP_ERROR.value, {
+                EVENT_ID: NPG_RESP_ERROR.value,
+              });
+              return {};
+            },
+            (myRes) => {
+              switch (myRes.status) {
+                case 200:
+                  mixpanel.track(NPG_SUCCESS.value, {
+                    EVENT_ID: NPG_SUCCESS.value,
+                  });
+                  pipe(
+                    myRes.value.paymentMethodData.form,
+                    validateSessionWalletCardFormFields,
+                    O.match(
+                      () => onError(NPG_RESP_ERROR.value),
+                      () => onResponse(myRes.value)
+                    )
+                  );
+                  return myRes;
+                case 401:
+                  mixpanel.track(NPG_RESP_ERROR.value, {
+                    EVENT_ID: NPG_RESP_ERROR.value,
+                  });
+                  onError(ErrorsType.UNAUTHORIZED);
+                  return {};
+                default:
+                  mixpanel.track(NPG_RESP_ERROR.value, {
+                    EVENT_ID: NPG_RESP_ERROR.value,
+                  });
+                  return {};
+              }
+            }
+          )
+        )
+    )
+  )();
+
+export const getFees = (
+  onSuccess: (value: boolean) => void,
+  onPspNotFound: () => void,
+  onError: (m: string) => void,
+  bin?: string
+) =>
+  calculateFees({
+    paymentId:
+      (getSessionItem(SessionItems.paymentMethod) as PaymentMethod | undefined)
+        ?.paymentMethodId || "",
+    bin,
+    onError,
+    onPspNotFound,
+    onResponsePsp: (resp) => {
+      pipe(
+        resp,
+        CalculateFeeResponse.decode,
+        O.fromEither,
+        O.chain((resp) => O.fromNullable(resp.belowThreshold)),
+        O.fold(
+          () => onError(ErrorsType.GENERIC_ERROR),
+          (value) => {
+            const firstPsp = pipe(
+              resp?.bundles,
+              O.fromNullable,
+              O.chain((sortedArray) => O.fromNullable(sortedArray[0])),
+              O.map((a) => a as Bundle),
+              O.getOrElseW(() => ({}))
+            );
+
+            setSessionItem(SessionItems.pspSelected, firstPsp);
+            onSuccess(value);
+          }
+        )
+      );
+    },
+  });
