+++ src/utils/api/helpers/ecommerce/nodeFaultCode.ts
@@ -1,0 +1,16 @@
+import * as t from "io-ts";
+
+const NodeFaultCodeR = t.interface({
+  faultCodeCategory: t.string,
+});
+
+const NodeFaultCodeO = t.partial({
+  faultCodeDetail: t.string,
+});
+
+export const NodeFaultCode = t.intersection(
+  [NodeFaultCodeO, NodeFaultCodeR],
+  "NodeFaultCode"
+);
+
+export type NodeFaultCode = t.TypeOf<typeof NodeFaultCode>;
